better code to calculate severity reward using useSeverityReward custom hook
