+++ src/components/Vulnerability/TermsAndProcess/TermsAndProcess.tsx
@@ -1,25 +1,18 @@
-import { useContext, useMemo, useState } from "react";
+import { useContext, useState } from "react";
 import { AccordionContext } from "../VulnerabilityAccordion";
 import { useTranslation } from "react-i18next";
-import { Colors, RC_TOOLTIP_OVERLAY_INNER_STYLE, TERMS_OF_USE } from "../../../constants/constants";
-import { ISeverity, IVault, IVaultDescription } from "../../../types/types";
-import { calculateRewardPrice, formatNumber } from "../../../utils";
-import Tooltip from "rc-tooltip";
-import InfoIcon from "../../../assets/icons/info.icon";
-import NFTMedia from "../../NFTMedia";
-import ZoomIcon from "../../../assets/icons/zoom.icon";
+import { TERMS_OF_USE } from "../../../constants/constants";
+import { IVault } from "../../../types/types";
 import { useSelector } from "react-redux";
 import { RootState } from "../../../reducers";
-import Modal from "../../Shared/Modal";
-import NFTPrize from "../../NFTPrize";
 import QuestionIcon from "../../../assets/icons/question-icon.svg";
 import ErrosIcon from "../../../assets/icons/error-icon.svg";
 import TimelineIcon from "../../../assets/icons/timeline-icon.svg";
 import BugIcon from "../../../assets/icons/bug-icon.svg";
 import RewardIcon from "../../../assets/icons/reward-icon.svg";
 import { Card, IVulnerabilityData } from "../types";
+import Award from "./components/Award/ Award";
 import "./index.scss";
-import { useVaultsTotalPrices } from "components/Vault/useVaultsTotalPrices";
 
 interface IProps {
   cards: IVulnerabilityData
@@ -28,41 +21,15 @@ interface IProps {
 export default function TermsAndProcess({ cards }: IProps) {
   const { submitCard } = useContext(AccordionContext);
   const [acceptedTermsOfUse, setAcceptedTermsOfUse] = useState(cards.terms.verified);
-  const [showNFTModal, setShowNFTModal] = useState(false);
-  const [modalNFTData, setModalNFTData] = useState<ISeverity | null>(null);
   const { t } = useTranslation();
 
   const projectId = cards.project.projectId;
   const vault = useSelector((state: RootState) => state.dataReducer.vaults!.filter((vault: IVault) => vault.id === projectId)[0]);
   const description = vault && vault.description;
-  const { tokenPrices } = useSelector((state: RootState) => state.dataReducer);
-  const { totalPrices } = useVaultsTotalPrices(vault.multipleVaults ?? [vault]);
-  const sumTotalPrices = Object.values(totalPrices).reduce((a, b = 0) => a + b, 0);
 
-  const prizes = useMemo(() => (description as IVaultDescription)?.severities?.map((severity: ISeverity, index: number) => {
-    const rewardPercentage = (Number(vault.rewardsLevels[severity.index]) / 10000) * 100;
-    const tokenPrice = tokenPrices?.[vault.stakingToken];
-    const rewardPrice = tokenPrice ? calculateRewardPrice(vault, rewardPercentage, tokenPrice, sumTotalPrices) : 0;
-
-    return (
-      <tr key={index} className={`${severity.name.toLocaleLowerCase().split(' ').join('-')}`}>
-        <td>{severity.name.toUpperCase()}</td>
-        <td className="total-vault-wrapper">
-          <b>{`${rewardPercentage.toFixed(2)}%`}</b> of vault &#8776; {`$${formatNumber(rewardPrice)}`}&nbsp;
-          <Tooltip
-            overlay="Prizes are in correlation to the funds in the vault and may change at any time"
-            overlayClassName="tooltip"
-            overlayInnerStyle={RC_TOOLTIP_OVERLAY_INNER_STYLE}>
-            <span><InfoIcon width="15" height="15" fill={Colors.white} /></span>
-          </Tooltip>
-        </td>
-        <td className="nft-wrapper" onClick={() => { setShowNFTModal(true); setModalNFTData(severity); }}>
-          <NFTMedia link={severity?.["nft-metadata"]?.image} width="50px" />
-          <ZoomIcon />
-        </td>
-      </tr>
-    )
-  }), [vault, description, tokenPrices, sumTotalPrices]);
+  const awards = description?.severities.map((severity, index) => {
+    return <Award key={index} vault={vault} severity={severity} />
+  })
 
   return (
     <div className="terms-and-process-wrapper">
@@ -113,7 +80,7 @@ export default function TermsAndProcess({ cards }: IProps) {
                 <th>Prize</th>
                 <th>NFT</th>
               </tr>
-              {prizes}
+              {awards}
             </tbody>
           </table>) : "Please choose project to view prizes"}
         {t("SubmitVulnerability.TermsAndProcess.awards-text-2")}
@@ -128,12 +95,6 @@ export default function TermsAndProcess({ cards }: IProps) {
         </div>
         <button disabled={!acceptedTermsOfUse} onClick={() => submitCard(Card.terms, Card.submission)}>NEXT</button>
       </div>
-      {
-        showNFTModal &&
-        <Modal title="NFT PRIZE" setShowModal={setShowNFTModal} maxWidth="600px" width="60%" height="fit-content">
-          <NFTPrize data={modalNFTData as any} />
-        </Modal>
-      }
     </div>
   )
 }
